+++ pyteal/__init__.pyi
@@ -3,6 +3,7 @@
 
 from pyteal.ast import *
 from pyteal.ast import __all__ as ast_all
+from pyteal.ast.pragma import pragma
 from pyteal.ir import *
 from pyteal.ir import __all__ as ir_all
 from pyteal.compiler import (
@@ -10,7 +11,6 @@ from pyteal.compiler import (
     MIN_TEAL_VERSION,
     DEFAULT_TEAL_VERSION,
     CompileOptions,
-    pragma,
     compileTeal,
     OptimizeOptions,
 )
@@ -20,6 +20,7 @@ from pyteal.errors import (
     TealTypeError,
     TealInputError,
     TealCompileError,
+    TealPragmaError,
 )
 from pyteal.config import MAX_GROUP_SIZE, NUM_SLOTS
 
@@ -138,6 +139,7 @@ __all__ = [
     "OptimizeOptions",
     "Or",
     "Pop",
+    "Pragma",
     "Reject",
     "Return",
     "ScratchIndex",
@@ -169,6 +171,7 @@ __all__ = [
     "TealInternalError",
     "TealLabel",
     "TealOp",
+    "TealPragmaError",
     "TealSimpleBlock",
     "TealType",
     "TealTypeError",
